requirements.txt
Django
django-el-pagination
django-mail-queue
django-registration-redux
django-versatileimagefield
djangorestframework
djangorestframework-simplejwt
Pillow
psycopg2-binary
PyJWT
python-decouple
python-magic
django-import-export
